instruct k8s to send us watch bookmarks
Watch bookmarks periodically sends updated resource versions for watches.
We already expect these in case we need to restart a failed watch. This commit
adds the required query option to the API call, fixing errors when restarting
the watch with a too old resource version.

Also fixes an issue that would restrict the updates we receive based on the
Pod selector even for non-pod resources.