pod evict: query k8s API for best eviction method
The eviction helper would fall back to "worse" eviction methods if the API
server reported the resource as missing. This is a problem when the reason
the resource was missing is not that method does not exist, but that the
resource was already deleted.

This meant that over time the ha-controller would gradually fall back
to the last supported method, even though better methods are available.

To fix this issue, we query the API server once at the start to list
the available methods, and then always select the one chosen.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>