failover: force delete pods on unresponsive nodes
Eviction always waits for confirmation by the kubelet to complete. This is
not ideal if the whole node crashed, as then kubelet can't confirm deletion,
and we are stuck with Pods in Terminating state. This is especially bad
in case of StatefulSets, as those won't ever recreate the Pod, as the old
one is still visible, even if it is factually gone.

The solution is to check the node state in the fail over case, and
if kubernetes reports that the node is not ready, we force delete the
pod from the API without waiting for any kind of acknowledgement by
the kubelet.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>