drbd: only parse needed items
The ad-hoc parser for the output of drbdsetup status would sometimes
fail. This was caused by large values in fields that only expected "int".

Instead of manually checking the possible range, we can just remove those
fields, as they do not hold any relevant information for the inner workings
of the HA Controller.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>