failover: fix safe volume detection using the wrong indexer
We try to detect the volumes of a Pod to determine if it safe to fail over.
However, we used the wrong index to get the persistent volume from the
configured persistent volume claim. This meant that Pods using PVCs
where considered unsafe, even if they where using DRBD resources.

To fix the issue, use the correct indexer name defined in a const.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>