skip volumes with no quorum configuration
We assumed that having no quorum configuration meant that the output of
"drbdsetup status --json" would also show "quorum: false". This turns out to be
a false assumption.

So we now check the output of "drbdsetup show --json" in addition to status,
and use the observed configuration to determine if a volume is eligble for
quorum based failover.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>